web-app/requirements.txt
amqp==2.1.4
anyjson==0.3.3
billiard==3.5.0.2
celery==4.0.2
Django==1.10.5
django-appconf==1.0.1
django-celery==3.1.17
django-channels==0.7.0
django-ckeditor==5.2.1
django-environ==0.4.4
django-image-cropping==1.0.4
django-taggit==0.21.3
django-tinymce==2.4.0
easy-thumbnails==2.3
gunicorn==19.6.0
kombu==4.0.2
oauthlib==2.0.1
olefile==0.44
Pillow==4.0.0
psycopg2==2.6.2
pytz==2016.10
redis==2.10.5
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
sqlparse==0.2.2
vine==1.1.3
